some_projects/aston_bank_transactions_etl/requirements.txt
Faker==33.1.0
numpy==2.1.3
pandas==2.2.3
psycopg2-binary==2.9.10
py4j==0.10.9.7
pyspark==3.5.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.2
